+++ plugins/kyoube-files/src/worker.ts
@@ -1,0 +1,7 @@
+import { runWorker } from "@paperclipai/plugin-sdk";
+import { createFilesPlugin } from "./plugin.js";
+
+const plugin = createFilesPlugin({});
+
+export default plugin;
+runWorker(plugin, import.meta.url);